app/requirements.txt
appnope==0.1.0
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.7
django-extensions==2.2.9
django-filter==2.3.0
djangorestframework==3.11.0
entrypoints==0.3
importlib-metadata==1.7.0
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
selenium==3.141.0
Send2Trash==1.5.0
Serializer==0.2.1
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.1.0
